+++ pages/main/chatbot/Store.js
@@ -1,0 +1,16 @@
+class Store {
+  constructor() {
+    this.state = {
+      nextReqGroup: 0,
+      prevTime: '',
+    };
+  }
+  getState(key) {
+    return this.state[key];
+  }
+  setState(key, newState) {
+    this.state[key] = newState;
+  }
+}
+
+export const store = new Store();
