+++ pages/main/index.html
@@ -9,6 +9,7 @@
     <link rel="stylesheet" href="/components/footer/footer.css" />
     <link rel="stylesheet" href="./chatbot/chatbot.css" />
     <script type="module" defer src="./chatbot/chatbotData.js"></script>
+    <script type="module" defer src="./chatbot/Store.js"></script>
     <script type="module" defer src="./chatbot/chatbotFunctions.js"></script>
     <script type="module" defer src="./chatbot/chatbot.js"></script>
     <script type="module" src="/components/component.js"></script>
